requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
environs==11.0.0
Markdown==3.7
marshmallow==3.22.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
pycparser==2.22
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2
